Grouparoo can use SQLite (#965)
* do not use findOrCreate - ever

* rollback team creation if something goes wrong

* move from custom queries to 2-phase updates

* Sqlite confiuration and aggregation changes

* Do not await logged models

* all non-null columns need default value for SQLite

* conditionally consider update return values (postgres only)

* Hooks which manipulate models use transaction if present

* ignore journal files

* don't set setFailedCount if page didn't load

* provide default string for JSON model parsing

* rollback lock usage

* update export response check

* do not use iLike for SQLite

* CLI generates .env to use SQLlite

* Staging should still use Postgres by default

* statically export PermissionTopics

* update pnpm lockfile

* model tests pass with SQLite

* Test SQLite in CI against models

* maxWorkers for sqlite tests

+++ core/api/src/models/Schedule.ts
@@ -265,22 +265,19 @@ export class Schedule extends LoggedModel<Schedule> {
   }
 
   @AfterDestroy
-  static async destroyAppOptions(instance: Schedule) {
-    return Option.destroy({
-      where: {
-        ownerGuid: instance.guid,
-      },
-    });
+  static async destroyAppOptions(instance: Schedule, { transaction }) {
+    return Option.destroy({ where: { ownerGuid: instance.guid }, transaction });
   }
 
   @AfterDestroy
-  static async stopRuns(instance: Schedule) {
+  static async stopRuns(instance: Schedule, { transaction }) {
     const runs = await Run.findAll({
       where: { creatorGuid: instance.guid, state: "running" },
+      transaction,
     });
 
     for (const i in runs) {
-      await runs[i].update({ state: "stopped" });
+      await runs[i].update({ state: "stopped" }, { transaction });
     }
   }
 }
